Moved Search to new directory structure
